fix verification and add script
